MAJOR REWORK: Repush: Make use of placeholders
Due to multiple people complaining about the required time for a document to appear after
being pushed, we now make use of "placeholders", which are simply blank files that pretend to be documents.

The purpose of placeholders is to create a xochtil document entry as fast as possible. When utilizing the
WebUI to transfer documents, the thumbnail of the document entry is created locally, meaning the greater the
resoltuion of the document, the more time xochitl will require to render a thumbnail and the longer it takes
for the document to appear on your device. To circuimvent this flaw, we push the palceholder file which will take little
to no time for it to appear on the device, and then we substitute the palceholder file with the actual document via scp.
This way, we only utilize the WebUI to create a fs entry, but transfer the actual document via scp.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>